moved importRepo method to importHelper
